Move lightning modules
